+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # BC Cache Changelog
 
-## Upcoming version 3.1.0 (????-??-??)
+## Upcoming version 3.1.0 (2023-01-04)
 
 ### Changed